+++ go.mod
@@ -11,4 +11,5 @@ require (
 	golang.org/x/mod v0.4.2
 	golang.org/x/tools v0.1.0
 	gopkg.in/yaml.v2 v2.4.0
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c
 )
