+++ internal/config/config.go
@@ -1,10 +1,5 @@
 package config
 
-import (
-	"gopkg.in/yaml.v2"
-	"os"
-)
-
 type DependenciesRule struct {
 	Package                     string   `yaml:"package"`
 	ShouldOnlyDependsOn         []string `yaml:"shouldOnlyDependsOn"`
@@ -46,18 +41,3 @@ type Config struct {
 	NamingRules       []*NamingRule       `yaml:"namingRules"`
 }
 
-func LoadConfig(configPath string) (*Config, error) {
-	config := &Config{}
-	file, err := os.Open(configPath)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	d := yaml.NewDecoder(file)
-	if err := d.Decode(&config); err != nil {
-		return nil, err
-	}
-
-	return config, nil
-}
