+++ internal/report/console/generator_test.go
@@ -1,0 +1,1 @@
+package console_test
