+++ internal/report/html/generator_test.go
@@ -1,0 +1,1 @@
+package html_test
